moves api.js into utils folder and corrects space between votes and error message, changes title in index.html to BBB
